feat: Verify signature before final frame
Resolves: #390

Data MUST be authenticated before it is released.
But there is a unique edge case
In the end of signed streams.
If the final frame is sent
but the signature has not been sent
the stream will release the plaintext
because the authTag has been received.
However the signature
has still not been verified.

This could lead to customers processing
data that is not valid
and circumvent one of the values of the signature suites
i.e. someone with decrypt can not forage
a new message.